chore(cameras): fix imports renamed folder + add camera configs in utility scripts
