+++ lerobot/find_cameras.py
@@ -36,10 +36,10 @@ import numpy as np
 from PIL import Image
 
 from lerobot.common.cameras.configs import ColorMode
-from lerobot.common.cameras.intel.camera_realsense import RealSenseCamera
-from lerobot.common.cameras.intel.configuration_realsense import RealSenseCameraConfig
 from lerobot.common.cameras.opencv.camera_opencv import OpenCVCamera
 from lerobot.common.cameras.opencv.configuration_opencv import OpenCVCameraConfig
+from lerobot.common.cameras.realsense.camera_realsense import RealSenseCamera
+from lerobot.common.cameras.realsense.configuration_realsense import RealSenseCameraConfig
 
 logger = logging.getLogger(__name__)
